Created unified buffer unit test file
