+++ apps/hardware_benchmarks/tests/merged_unit_tests/test_main.cpp
@@ -11,17 +11,12 @@
 
 #include <fstream>
 #include "test_utils.h"
-
-#define PRINT_PASSED(msg) std::cout << GREEN << msg << " test passed." << RESET << std::endl;
+#include "ubuffer_tests.h"
 
 using namespace CoreIR;
 using namespace Halide;
 using namespace Halide::Tools;
 using namespace std;
-
-std::string GREEN = "\033[32m";
-std::string RED = "\033[31m";
-std::string RESET = "\033[0m";
 
 template<typename T>
 bool is2D(T& buf) {
@@ -152,52 +147,6 @@ void runHWKernel(const std::string& inputName, CoreIR::Module* m, Halide::Runtim
 template<typename T>
 void runHWKernel(CoreIR::Module* m, Halide::Runtime::Buffer<T>& hwInputBuf, Halide::Runtime::Buffer<T>& outputBuf) {
   runHWKernel("self.in_arg_0_0_0", m, hwInputBuf, outputBuf);
-}
-
-CoreIR::Context* hwContext() {
-  CoreIR::Context* context = newContext();
-  CoreIRLoadLibrary_commonlib(context);
-  CoreIRLoadLibrary_float(context);
-  return context;
-}
-
-CoreIR::Module* buildModule(Halide::Internal::HardwareInfo& info, bool useUbuffer, CoreIR::Context* context, const std::string& name, std::vector<Argument>& args, const std::string& fName, Func& hwOutput) {
-  Target t;
-  t = t.with_feature(Target::Feature::CoreIR);
-  if (!useUbuffer) {
-    //t = t.with_feature(Target::Feature::UseExtractHWKernel);
-  }
-  auto hm = hwOutput.compile_to_module(args, name, t);
-  cout << "Compiled to module..." << endl;
-  cout << hm << endl;
-  for (auto f : hm.functions()) {
-    cout << "Generating coreir for function " << f.name << endl;
-    Halide::Internal::CodeGen_CoreHLS_Kernel gen("conv_3_3_app.json");
-    gen.info = info;
-    f.body.accept(&gen);
-  }
-
-  if (!loadFromFile(context, "./conv_3_3_app.json")) {
-    cout << "Error: Could not load json for unit test!" << endl;
-    context->die();
-  }
-  context->runPasses({"rungenerators", "flattentypes", "flatten", "wireclocks-coreir"});
-  CoreIR::Module* m = context->getNamespace("global")->getModule("DesignTop");
-  cout << "Module after wiring clocks ..." << endl;
-  m->print();
-  return m;
-}
-
-CoreIR::Module* buildModule(const bool useUbuffer, CoreIR::Context* context, const std::string& name, std::vector<Argument>& args, const std::string& fName, Func& hwOutput) {
-  Halide::Internal::HardwareInfo info;
-  info.hasCriticalPathTarget = false;
-  return buildModule(info, useUbuffer, context, name, args, fName, hwOutput);
-}
-
-CoreIR::Module* buildModule(CoreIR::Context* context, const std::string& name, std::vector<Argument>& args, const std::string& fName, Func& hwOutput) {
-  Halide::Internal::HardwareInfo info;
-  info.hasCriticalPathTarget = false;
-  return buildModule(info, false, context, name, args, fName, hwOutput);
 }
 
 template<typename T>
@@ -1713,119 +1662,6 @@ void small_conv_3_3_critical_path_test() {
   deleteContext(context);
  
   PRINT_PASSED("Conv 3x3 with critical path test passed");
-}
-
-void ubuffer_small_conv_3_3_test() {
-  ImageParam input(type_of<uint8_t>(), 2);
-  Func output;
-
-  Var x("x"), y("y");
-
-  Func kernel("kernel");
-  Func conv("conv");
-  RDom r(0, 3,
-      0, 3);
-
-  kernel(x,y) = 0;
-  kernel(0,0) = 11;      kernel(0,1) = 12;      kernel(0,2) = 13;
-  kernel(1,0) = 14;      kernel(1,1) = 0;       kernel(1,2) = 16;
-  kernel(2,0) = 17;      kernel(2,1) = 18;      kernel(2,2) = 19;
-
-  conv(x, y) = 0;
-
-  Func hw_input("hw_input");
-  hw_input(x, y) = cast<uint16_t>(input(x, y));
-  conv(x, y)  += kernel(r.x, r.y) * hw_input(x + r.x, y + r.y);
-
-  Func hw_output("hw_output");
-  hw_output(x, y) = cast<uint8_t>(conv(x, y));
-  output(x, y) = hw_output(x,y);
-
-  Var xi,yi, xo,yo;
-
-  hw_input.compute_root();
-  hw_output.compute_root();
-
-  int inTileSize = 4;
-  int outTileSize = inTileSize - 2;
-
-  hw_output.bound(x, 0, outTileSize);
-  hw_output.bound(y, 0, outTileSize);
-
-  output.bound(x, 0, outTileSize);
-  output.bound(y, 0, outTileSize);
-
-  // Creating input data
-  Halide::Buffer<uint8_t> inputBuf(4, 4);
-  Halide::Runtime::Buffer<uint8_t> hwInputBuf(4, 4, 1);
-  for (int i = 0; i < 4; i++) {
-    for (int j = 0; j < 4; j++) {
-      for (int b = 0; b < 1; b++) {
-        inputBuf(i, j, b) = i + j*2;
-        hwInputBuf(i, j, b) = inputBuf(i, j, b);
-      }
-    }
-  }
- 
-  // Creating CPU reference output
-  Halide::Buffer<uint8_t> cpuOutput(2, 2);
-  ParamMap rParams;
-  rParams.set(input, inputBuf);
-  Target t;
-  hw_output.realize(cpuOutput, t, rParams);
-  
-  Halide::Runtime::Buffer<uint8_t> outputBuf(2, 2, 1);
-  
-  int tileSize = 4;
-  hw_output.tile(x,y, xo,yo, xi,yi, tileSize-2, tileSize-2)
-    .hw_accelerate(xi, xo);
-
-  kernel.compute_at(hw_output, xo)
-    .unroll(x).unroll(y);
-
-  conv.update()
-    .unroll(r.x, 3)
-    .unroll(r.y, 3);
-  conv.linebuffer();
-
-  hw_input.stream_to_accelerator();
-
-  // Generate CoreIR
-  auto context = hwContext();
-  vector<Argument> args{input};
-  buildModule(context, "coreir_conv_3_3", args, "conv_3_3", hw_output);
-
-  // Get unified buffer fie
-  if (!loadFromFile(context, "./ubuffers.json")) {
-    cout << "Error: Could not load json for ubuffer test!" << endl;
-    context->die();
-  }
-  context->runPasses({"rungenerators", "flattentypes", "flatten", "wireclocks-coreir"});
-  CoreIR::Module* m = context->getNamespace("global")->getModule("hw_input_ubuffer");
-  cout << "hw_input_ubuffer..." << endl;
-  m->print();
-
-  SimulatorState state(m);
-  state.setValue("self.write_port_0", BitVector(16, 0));
-  state.setValue("self.write_port_0_en", BitVector(1, 0));
-  state.setClock("self.clk", 0, 1);
-  state.setValue("self.reset", BitVector(1, 1));
-  state.resetCircuit();
-  state.setValue("self.reset", BitVector(1, 0));
-
-  assert(state.getBitVec("self.read_port_0_valid") == BitVec(1, 0));
-
-  state.setValue("self.write_port_0", BitVector(16, 1));
-  state.setValue("self.write_port_0_en", BitVector(1, 1));
-  
-  state.execute();
-
-  assert(state.getBitVec("self.read_port_0_valid") == BitVec(1, 0));
-  
-  deleteContext(context);
- 
-  cout << GREEN << "UBuffer to linebuffer for conv 3x3 test passed" << RESET << endl;
-  assert(false);
 }
 
 void small_conv_3_3_test() {
