+++ apps/hardware_benchmarks/tests/merged_unit_tests/ubuffer_tests.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void ubuffer_small_conv_3_3_test();
